Change a way static plugins are handled
