+++ src/plugins/CMakeLists.txt
@@ -6,7 +6,6 @@ set(NOX_PLUGINS_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/plugin.h
     ${CMAKE_CURRENT_SOURCE_DIR}/plugin_loader.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/plugin_loader.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/static_plugin.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/static_plugin.h
 )
